[feat] : 삭제 시에 확인용 토스트를 띄우도록 수정, 요구사항 검색 시 프레임워크, 언어를 사용하도록 수정

+++ src/components/Requirement/RequirementControlBtn.tsx
@@ -1,14 +1,20 @@
 import { css } from '@emotion/react';
+import { useEffect } from 'react';
 import { toast } from 'react-toastify';
+import { useRecoilValue } from 'recoil';
 
+import deleteStateAtom from '@/atoms/deleteStateAtom';
+import { devEnvironmentAtom } from '@/atoms/devEnvironmentAtom';
 import { useRequirement } from '@/hooks/useRequirement';
 import { theme, ThemeType } from '@/styles/theme';
 import { RequirementStateType } from '@/types/requirement';
+import { copyText, generateSearchSentence } from '@/utils/gpt';
 
 import HoverIcon from '../common/HoverIcon';
 import GPTGenerateCodeBtn from '../gpt/GPTGenerateCodeBtn';
 import DeleteIcon from '../icons/DeleteIcon';
 import SearchIcon from '../icons/SearchIcon';
+import DeleteConfirmation from '../Toast/DeleteConfirmation';
 
 type RequirementControlBtnProps = {
   requirement: RequirementStateType;
@@ -19,29 +25,43 @@ export default function RequirementControlBtn({
   requirement,
   onSelectId,
 }: RequirementControlBtnProps) {
+  const devState = useRecoilValue(devEnvironmentAtom);
+  const isDelete = useRecoilValue(deleteStateAtom(requirement.id));
+
   const { deleteRequire } = useRequirement();
 
-  const handleSearch = async () => {
-    try {
-      await navigator.clipboard.writeText(requirement.title);
-      toast.success('요구사항을 복사했어요. 붙여넣기로 GPT에 바로 검색해 보세요!');
-      setTimeout(() => {
-        window.open('https://chat.openai.com', '_blank');
-      }, 1200);
-    } catch (e) {
-      toast.error('요구사항 복사에 실패했어요. 다시 시도해 주세요.');
-    }
+  const handleSearch = () => {
+    const text = generateSearchSentence({
+      framework: devState.framework,
+      language: devState.language,
+      title: requirement.title,
+    });
+
+    copyText({
+      text,
+      toastMessage: '요구사항을 복사했어요!',
+      isRequirement: true,
+    });
   };
 
   const handleDelete = () => {
-    toast.warn('정말 삭제하시겠습니까?');
-    deleteRequire({ id: requirement.id });
+    toast.warn(<DeleteConfirmation id={requirement.id} />);
   };
+
+  useEffect(() => {
+    if (isDelete) {
+      deleteRequire({ id: requirement.id });
+    }
+  }, [isDelete]);
 
   return (
     <div css={requirementItemBtnStyle(theme)}>
       <HoverIcon icon={<GPTGenerateCodeBtn requirement={requirement} onSelectId={onSelectId} />} />
-      <HoverIcon icon={<SearchIcon size={18} />} onClick={handleSearch} />
+      <HoverIcon
+        className="requirement-search-btn"
+        icon={<SearchIcon size={18} />}
+        onClick={handleSearch}
+      />
       <HoverIcon
         className="requirement-delete-btn"
         icon={<DeleteIcon size={18} />}
@@ -57,8 +77,11 @@ const requirementItemBtnStyle = (theme: ThemeType) => css`
   padding: 2px 5px;
   color: ${theme.colors.green};
 
+  .requirement-search-btn {
+    color: ${theme.colors.beige};
+  }
+
   .requirement-delete-btn {
-    color: ${theme.colors.red};
     opacity: 0.2;
     :hover {
       opacity: 0.6;
